refactor(task): split task.go into domain-focused handler files
Decompose the 1953-line task.go into four focused files:
- task_handlers.go: CRUD, field updates, bulk ops, activation (+ Handler struct, shared types)
- template_handlers.go: template/recurring handlers
- query_handlers.go: structured and date-based query handlers
- nlp_handlers.go: natural language AI handlers and applyEditInstructions helper

Pure file move — no logic changes. All tests pass.
